Fixed a bug in the file rollback. When rolling back the enlistment was trying to delete files that may not exist, throwing an exception and breaking the transaction. Now, if the file isn't there the enlistment doesn't try to delete it as the file system is already in the desired state.

+++ ChinhDo.Transactions.FileManager/TxEnlistment.cs
@@ -286,7 +286,8 @@ namespace ChinhDo.Transactions
 					}
 					else
 					{
-						File.Delete(_originalFileName);
+						if (File.Exists(_originalFileName))
+							File.Delete(_originalFileName);
 					}
 				}
